LRQA-75167 Move test into CFG testing file since we are testing CFG functionality
